Split out non-class methods into utils module.
